+++ client/src/pages/LoginPage.jsx
@@ -41,7 +41,7 @@ export default function LoginPage() {
         <p style={{ textAlign: 'center', marginTop: '20px', fontSize: '0.8rem', color: 'var(--text3)' }}>
           L'accès est sur invitation uniquement.<br />Contactez un administrateur pour obtenir un compte.
         </p>
-        <p style={{ textAlign: 'center', marginTop: '32px', fontSize: '0.7rem', color: 'var(--text3)', opacity: 0.7 }}>v1.2.0</p>
+        <p style={{ textAlign: 'center', marginTop: '32px', fontSize: '0.7rem', color: 'var(--text3)', opacity: 0.7 }}>v1.2.1</p>
       </div>
     </div>
   )